SREP-963: Change to use Machine API
The previous cluster-config-v1 configmap is deprecated and the interim
suggestion is to use the Machine api (until something more mature is
developed). To that end, I have abstracted somewhat the way in which the
data is accessed. Instead of being fixed to a configmap, the user may
instead pass whatever kubectl fragment (eg `get machines`) as well as
the `yq` expressions that match, so that when the next data source(s)
are available, the transition will be easier.

Signed-off-by: Lisa Seelye <lseelye@redhat.com>